Add ability to delete Groups

+++ src/components/home/GroupsView.js
@@ -44,8 +44,10 @@ const GroupsView = ({
 
   const handleDelete = () => {
     if (active.group) {
+      const confirm = window.confirm(`Delete Group ${active.group.name}?`);
+      if (!confirm) return;
       deleteGroup(active.group.id, active.group.name);
-    }
+    };
   };
 
   const renderToolbar = () => {
